Completely rework registry import
vscode-windows-registry choked on WOW6432Node; replacing with node-cmd calls to reg query and caching (on my machine with 70+ server definitions it was VERY slow to run 140+ commands individually; with the caching it boils down to just 5).

Also includes usability improvements as requested in review comments.

Note: on my machine, I see data both there and in the original place I was looking, and there seems to be preference for WOW6432Node, so I let that override the other location.